Добавляет слайс app
