+++ project/src/store/actions.ts
@@ -1,0 +1,7 @@
+import { createAction } from '@reduxjs/toolkit';
+import { AppRoute } from '../const';
+
+export const redirectToRoute = createAction(
+  'app/redirectToRoute',
+  (route: AppRoute) => ({payload: route})
+);
