--- project/src/store/app/actions.ts
@@ -1,13 +1,0 @@
-import { createAction } from '@reduxjs/toolkit';
-
-import { AppRoute } from '../../const';
-
-export const redirectToRoute = createAction(
-  'app/redirectToRoute',
-  (route: AppRoute) => ({payload: route})
-);
-
-export const setHeaderNavigationStatus = createAction(
-  'app/setHeaderNavigationStatus',
-  (hasHeaderNavigation: boolean) => ({payload: hasHeaderNavigation})
-);
